+++ products/tracing/backend/logic.py
@@ -5,9 +5,6 @@ Validation, calculations, business rules, ORM queries.
 Called by facade/api.py.
 """
 
-import json
-import base64
-import decimal
 import datetime as dt
 from typing import TYPE_CHECKING
 from zoneinfo import ZoneInfo
@@ -18,8 +15,6 @@ from posthog.schema import (
     HogQLFilters,
     IntervalType,
     PropertyGroupsMode,
-    SpanPropertyFilter,
-    SpanPropertyFilterType,
     TraceSpansQuery,
     TraceSpansQueryResponse,
 )
@@ -27,7 +22,6 @@ from posthog.schema import (
 from posthog.hogql import ast
 from posthog.hogql.constants import HogQLGlobalSettings, LimitContext
 from posthog.hogql.parser import parse_expr, parse_order_expr, parse_select
-from posthog.hogql.property import property_to_expr
 from posthog.hogql.query import execute_hogql_query
 
 from posthog.clickhouse.client.connection import Workload
@@ -36,45 +30,16 @@ from posthog.hogql_queries.query_runner import AnalyticsQueryRunner, QueryRunner
 from posthog.hogql_queries.utils.query_date_range import QueryDateRange
 from posthog.models.filters.mixins.utils import cached_property
 
+from products.tracing.backend.constants import TRACE_SPANS_LIST_SETTINGS
+from products.tracing.backend.filter_builder import TraceSpansFilterBuilder
+from products.tracing.backend.query_date_range import tracing_qdr_baseline, tracing_qdr_minutely
+
 if TYPE_CHECKING:
     from posthog.models import Team, User
 
 
-def _normalise_to_base64(value: str) -> str:
-    try:
-        int(value, 16)
-        return base64.b64encode(bytes.fromhex(value)).decode()
-    except ValueError:
-        return value
-
-
-def _is_number(value: str) -> bool:
-    try:
-        float(value)
-        return True
-    except ValueError:
-        return False
-
-
-# OTel span.kind enum labels sent from the filter UI, mapped to their wire integer values.
-_SPAN_KIND_LABEL_TO_INT: dict[str, int] = {
-    "Unspecified": 0,
-    "Internal": 1,
-    "Server": 2,
-    "Client": 3,
-    "Producer": 4,
-    "Consumer": 5,
-}
-
-# OTel status_code label → int. 'OK' expands to {0, 1} so 'unset' spans match 'ok' filters.
-_STATUS_CODE_LABEL_TO_INTS: dict[str, list[int]] = {
-    "OK": [0, 1],
-    "Error": [2],
-}
-
-
 class TraceSpansQueryRunnerMixin(QueryRunner):
-    """Shared WHERE clause and settings for all trace span query runners."""
+    """Shared date range, filter builder, and settings for trace span query runners."""
 
     def __init__(self, query: TraceSpansQuery, *args, **kwargs) -> None:
         super().__init__(query, *args, **kwargs)
@@ -87,164 +52,10 @@ class TraceSpansQueryRunnerMixin(QueryRunner):
 
         self.modifiers.convertToProjectTimezone = False
         self.modifiers.propertyGroupsMode = PropertyGroupsMode.OPTIMIZED
-
-        def get_property_type(value: str | float | bool) -> str:
-            try:
-                float(value)
-                return "float"
-            except (ValueError, TypeError):
-                pass
-            return "str"
-
-        self.span_filters: list[SpanPropertyFilter] = []
-        self.span_attribute_filters: list[SpanPropertyFilter] = []
-        self.resource_attribute_filters: list[SpanPropertyFilter] = []
-        if self.query.filterGroup and self.query.filterGroup.values:
-            for property_group in self.query.filterGroup.values:
-                for prop in property_group.values:
-                    prop_type = getattr(prop, "type", None)
-                    if prop_type == SpanPropertyFilterType.SPAN_RESOURCE_ATTRIBUTE:
-                        self.resource_attribute_filters.append(prop)
-                    if prop_type == SpanPropertyFilterType.SPAN:
-                        self.span_filters.append(prop)
-                    elif prop_type == SpanPropertyFilterType.SPAN_ATTRIBUTE:
-                        if isinstance(prop, SpanPropertyFilter) and prop.value:
-                            property_type = "str"
-                            if isinstance(prop.value, list):
-                                property_types = {get_property_type(v) for v in prop.value}
-                                if len(property_types) == 1:
-                                    property_type = property_types.pop()
-                            else:
-                                property_type = get_property_type(prop.value)
-
-                            prop = prop.model_copy(deep=True)
-                            prop.key = f"{prop.key}__{property_type}"
-
-                        self.span_attribute_filters.append(prop)
+        self._filter_builder = TraceSpansFilterBuilder(self.team, self.query)
 
     def where(self) -> ast.Expr:
-        exprs: list[ast.Expr] = []
-
-        exprs.append(
-            parse_expr(
-                "toStartOfDay(time_bucket) >= toStartOfDay({date_from}) and toStartOfDay(time_bucket) <= toStartOfDay({date_to})",
-                placeholders={
-                    **self.query_date_range.to_placeholders(),
-                },
-            )
-        )
-
-        exprs.append(ast.Placeholder(expr=ast.Field(chain=["filters"])))
-
-        if self.query.serviceNames:
-            exprs.append(
-                parse_expr(
-                    "service_name IN {serviceNames}",
-                    placeholders={
-                        "serviceNames": ast.Tuple(exprs=[ast.Constant(value=str(sn)) for sn in self.query.serviceNames])
-                    },
-                )
-            )
-
-        if self.query.statusCodes:
-            exprs.append(
-                parse_expr(
-                    "status_code IN {statusCodes}",
-                    placeholders={
-                        "statusCodes": ast.Tuple(exprs=[ast.Constant(value=int(sc)) for sc in self.query.statusCodes])
-                    },
-                )
-            )
-
-        if self.span_filters:
-            for span_filter in self.span_filters:
-                if span_filter.key in ("trace_id", "span_id"):
-                    if isinstance(span_filter.value, list):
-                        span_filter.value = [_normalise_to_base64(str(v)) for v in span_filter.value]
-                    else:
-                        span_filter.value = _normalise_to_base64(str(span_filter.value))
-
-                if span_filter.key in ("duration"):
-                    span_filter.key = "duration_nano"
-
-                    if isinstance(span_filter.value, list):
-                        span_filter.value = [
-                            str(decimal.Decimal(str(v)) * 1000000) for v in span_filter.value if _is_number(str(v))
-                        ]
-                    else:
-                        if _is_number(str(span_filter.value)):
-                            span_filter.value = str(decimal.Decimal(str(span_filter.value)) * 1000000)
-
-                # Filter UI stores human labels for kind/status_code so the applied-filter
-                # chip reads naturally. Translate labels to the integer column values here.
-                if span_filter.key == "kind":
-                    values = span_filter.value if isinstance(span_filter.value, list) else [str(span_filter.value)]
-                    span_filter.value = [
-                        _SPAN_KIND_LABEL_TO_INT[str(v)] for v in values if str(v) in _SPAN_KIND_LABEL_TO_INT
-                    ]
-
-                if span_filter.key == "status_code":
-                    values = span_filter.value if isinstance(span_filter.value, list) else [str(span_filter.value)]
-                    expanded: list[int] = []
-                    for v in values:
-                        if str(v) in _STATUS_CODE_LABEL_TO_INTS:
-                            expanded.extend(_STATUS_CODE_LABEL_TO_INTS[str(v)])
-                    span_filter.value = [str(v) for v in expanded]
-
-                exprs.append(property_to_expr(span_filter, team=self.team))
-
-        if self.span_attribute_filters:
-            exprs.append(property_to_expr(self.span_attribute_filters, team=self.team))
-
-        if self.resource_attribute_filters:
-            for f in self.resource_attribute_filters:
-                exprs.append(property_to_expr(f, team=self.team))
-
-        if self.query.traceId:
-            trace_id_b64 = base64.b64encode(bytes.fromhex(self.query.traceId)).decode("ascii")
-            exprs.append(
-                parse_expr(
-                    "trace_id = {traceId}",
-                    placeholders={
-                        "traceId": ast.Constant(value=trace_id_b64),
-                    },
-                )
-            )
-
-        if self.query.after:
-            try:
-                cursor = json.loads(base64.b64decode(self.query.after).decode("utf-8"))
-                cursor_ts = dt.datetime.fromisoformat(cursor["timestamp"])
-                cursor_uuid = cursor["uuid"]
-            except (KeyError, ValueError, json.JSONDecodeError) as e:
-                raise ValueError(f"Invalid cursor format: {e}")
-
-            op = ">" if self.query.orderBy == "earliest" else "<"
-            ts_op = ">=" if self.query.orderBy == "earliest" else "<="
-
-            exprs.append(
-                parse_expr(
-                    f"time_bucket {ts_op} toStartOfDay({{cursor_ts}})",
-                    placeholders={"cursor_ts": ast.Constant(value=cursor_ts)},
-                )
-            )
-            exprs.append(
-                parse_expr(
-                    f"timestamp {ts_op} {{cursor_ts}}",
-                    placeholders={"cursor_ts": ast.Constant(value=cursor_ts)},
-                )
-            )
-            exprs.append(
-                parse_expr(
-                    f"(timestamp, uuid) {op} ({{cursor_ts}}, {{cursor_uuid}})",
-                    placeholders={
-                        "cursor_ts": ast.Constant(value=cursor_ts),
-                        "cursor_uuid": ast.Constant(value=cursor_uuid),
-                    },
-                )
-            )
-
-        return ast.And(exprs=exprs)
+        return self._filter_builder.build_where(self.query_date_range)
 
     @cached_property
     def query_date_range(self) -> QueryDateRange:
@@ -290,13 +101,7 @@ class TraceSpansQueryRunnerMixin(QueryRunner):
 
     @cached_property
     def settings(self) -> HogQLGlobalSettings:
-        return HogQLGlobalSettings(
-            allow_experimental_object_type=False,
-            allow_experimental_join_condition=False,
-            transform_null_in=False,
-            max_bytes_to_read=None,
-            read_overflow_mode=None,
-        )
+        return TRACE_SPANS_LIST_SETTINGS
 
 
 class TraceSpansQueryRunner(TraceSpansQueryRunnerMixin, AnalyticsQueryRunner[TraceSpansQueryResponse]):
@@ -430,13 +235,7 @@ def run_service_names_query(
     search: str = "",
 ) -> list[dict]:
     """Return distinct service names from trace spans."""
-    query_date_range = QueryDateRange(
-        date_range=date_range,
-        team=team,
-        interval=IntervalType.MINUTE,
-        interval_count=2,
-        now=dt.datetime.now(),
-    )
+    query_date_range = tracing_qdr_minutely(team, date_range)
 
     exprs: list[ast.Expr] = [
         parse_expr(
@@ -493,14 +292,7 @@ def run_attribute_names_query(
     offset: int = 0,
 ) -> tuple[list[dict], int]:
     """Return attribute names from trace_attributes table."""
-    query_date_range = QueryDateRange(
-        date_range=date_range,
-        team=team,
-        interval=IntervalType.MINUTE,
-        interval_count=10,
-        now=dt.datetime.now(),
-        timezone_info=ZoneInfo("UTC"),
-    )
+    query_date_range = tracing_qdr_baseline(team, date_range)
 
     property_filter_type = (
         attribute_type if attribute_type in ("span", "span_attribute", "span_resource_attribute") else "span_attribute"
@@ -566,14 +358,7 @@ def run_attribute_values_query(
     offset: int = 0,
 ) -> list[dict]:
     """Return attribute values for a given key from trace_attributes table."""
-    query_date_range = QueryDateRange(
-        date_range=date_range,
-        team=team,
-        interval=IntervalType.MINUTE,
-        interval_count=10,
-        now=dt.datetime.now(),
-        timezone_info=ZoneInfo("UTC"),
-    )
+    query_date_range = tracing_qdr_baseline(team, date_range)
 
     query = parse_select(
         """
